feat: use common params for kv.Create
kv.Create method does not allow to pass the context and CreateOptions.
In order to align the custom method signature with the generated one
introduce the context and the CreateOptions.
This effort, once all the common methods' signatures are aligned with
the generated ones, will converge in the deletion of the
custom methods' implementations, embedding the generated interface.

Signed-off-by: fossedihelm <ffossemo@redhat.com>
(cherry picked from commit e844a59b6447720a4d626df32aea1e3f9c70ae92)