feat: use common params for kv.Update
kv.Update and kv.UpdateStatus methods do not allow to pass the context and UpdateOptions.
In order to align the custom methods signature with the generated ones introduce
the context and the UpdateOptions.
This effort, once all the common methods' signatures are aligned with
the generated ones, will converge in the deletion of the
custom methods' implementations, embedding the generated interface.

Signed-off-by: fossedihelm <ffossemo@redhat.com>
(cherry picked from commit d475ba2bb4d5ffdfe8ce99934ee7977d413bfadf)