feat: use common params for kv.Delete
kv.Delete method does not allow to pass the context and uses the pointer to DeleteOptions.
In order to align the custom method signature with the generated one
switch using the DeleteOptions instead of the pointer one and introduce the context.
This effort, once all the common methods' signatures are aligned with
the generated ones, will converge in the deletion of the
custom methods' implementations, embedding the generated interface.

Signed-off-by: fossedihelm <ffossemo@redhat.com>
(cherry picked from commit bdc20cab2e8504ae3451b0bb53ff47877a3ae652)