feat: embed generated client for KV
The embed will provide us new methods (`DeleteCollection` and `Watch`).
Our `KubeVirtInterface` is still needed, otherwise, we will not have
the PatchStatus method.
Also, generally, before dropping our interfaces we first need to
adjust all the unit tests to use the fakeclient.
This is because when we drop our interface we lose also the
mockClient (that is required at the moment for the uni tests).

Signed-off-by: fossedihelm <ffossemo@redhat.com>