feat: use common params for VMIP.Patch
vmip.Patch method does not allow to pass the context and PatchOptions.
In order to align the custom method signature with the generated one introduce
the context and the PatchOptions.
This effort, once all the common methods' signatures are aligned with
the generated ones, will converge in the deletion of the
custom methods' implementations, embedding the generated interface.

Signed-off-by: fossedihelm <ffossemo@redhat.com>
(cherry picked from commit 8dd0c888b6270e0888912aae7edd9597b53fb3bb)