feat: use common params for VMIP.Delete
vimp.Delete method does not allow to pass the context and uses the pointer to DeleteOptions.
In order to align the custom method signature with the generated one
switch using the DeleteOptions instead of the pointer one and introduce the context.
This effort, once all the common methods' signatures are aligned with
the generated ones, will converge in the deletion of the
custom methods' implementations, embedding the generated interface.

Signed-off-by: fossedihelm <ffossemo@redhat.com>
(cherry picked from commit 0448a50a77b407a4fae0ce0a4ea9780dcae94338)