feat: use common params for VMIRS.Create
rs.Create method does not allow to pass the context and CreateOptions.
In order to align the custom method signature with the generated one introduce
the context and the CreateOptions.
This effort, once all the common methods' signatures are aligned with
the generated ones, will converge in the deletion of the
custom methods' implementations, embedding the generated interface.

Signed-off-by: fossedihelm <ffossemo@redhat.com>

(cherry picked from commit ee14c4bdd122a3136a5a27d6ebb1ce5c2eed3146)
Signed-off-by: fossedihelm <ffossemo@redhat.com>