feat: use common params for VMIRS.Update
rs.Update and rs.UpdateStatus methods do not allow to pass the context and UpdateOptions.
In order to align the custom method signatures with the generated ones introduce
the context and the UpdateOptions.
This effort, once all the common methods' signatures are aligned with
the generated ones, will converge in the deletion of the
custom methods' implementations, embedding the generated interface.

Signed-off-by: fossedihelm <ffossemo@redhat.com>
(cherry picked from commit 4b0796d7723fb1f8e2bddd58d8a9fcfce492ebac)