feat: use common params for VMIRS.Patch
rs.Patch and rs.PatchStatus methods do not allow to pass the context and PatchOptions.
In order to align the custom method signatures with the generated ones introduce
the context and the PatchOptions.
This effort, once all the common methods' signatures are aligned with
the generated ones, will converge in the deletion of the
custom methods' implementations, embedding the generated interface.

Signed-off-by: fossedihelm <ffossemo@redhat.com>