feat: use common params for VMIRS.Get
rs.Get does not allow to pass the context.
In order to align the custom method signature with
the generated ones introduce the context.
This effort, once all the common methods' signatures are aligned with
the generated ones, will converge in the deletion of the
custom methods' implementations, embedding the generated interface.

Signed-off-by: fossedihelm <ffossemo@redhat.com>