cmd, cniplugins, passt-binding: Build the binary statically
The custom CNI plugin that is used with the `passt` network binding
plugin is distributed through the release assets.

In order to allow it to run on different platforms, build it statically
to avoid shared library conflicts.

This change should resolve issues [1] reported by users when attempting
to use the `passt` binding plugin.

[1] https://github.com/kubevirt/kubevirt/issues/11586

Signed-off-by: Edward Haas <edwardh@redhat.com>