Fix the monitoring e2e test (#2862)
* Fix the monitoring e2e test

The "KubeVirtCRModified alert should fired when there is a modification
on a CR" e2e test fails frequently, because the test logic is trying to
trigger the `KubeVirtCRModified` alert, by adding a new label to
KubeVirt CR, while we now allowing adding custom lables.

This PR adds a fake feature gate to the KubeVirt CR in order to trigger
the same alert. HCO fully reconciles the KV feature gate, so it should
trigger the alert in this case.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* use predefined e2e label

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

---------

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>