feature-gates: Retrospectively document CommonInstancetypesDeploymentGate as Beta from 1.2.0
With the feature lifecycle discussion [1] heading towards consensus this
change retrospectively moves the CommonInstancetypesDeploymentGate to
Beta from 1.2.0 with an eye on moving to GA and removal of the FG by 1.4.0.

This is done retrospectively as the FG itself is now enabled by default
by HCO with KubeVirt 1.2 and as such is already meeting the criteria of
the beta phase defined by the feature lifecycle document.

As this is purely documentation the impact of this change is zero and
just helps us move the process forward in the coming 1.3 and 1.4
releases.

[1] https://github.com/kubevirt/community/pull/251

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>