VMClone create mutator: Don't mutate when there is no need
Currently, the VMClone create mutator
creates a JSON patch to replace the admitted object's
spec even when there is no need to replace it (spec.target is fully set).

Do not add a JSON patch to the AdmissionResponse in case there
is no need to mutate the object (the target hasn't been changed).

Signed-off-by: Orel Misan <omisan@redhat.com>