net, admitter, test: Place a general cluster-config-checker
The unit tests have been using a cluster-config-checker stub that was
originally focused on SLIRP tests. Later it was used to just satisfy
other scenarios which had nothing to do with SLIRP.

As new cluster-config methods are needed for other scenarios, the
current stub is generalized by name and location.

Signed-off-by: Edward Haas <edwardh@redhat.com>