virt-api, admitter: Remove passt as single interface validation
The passt core binding has a validation that checks that it is defined
without any other interface in the spec.

We support only L2 bindings on the secondary interfaces which use
bridging to ling the pod interface to the VM. Therefore, there is no
real limitation to have additional interfaces to passt.

In addition, passt is targeted for removal after declared as deprecated
in favor of the binding plugin version of it.

Therefore, this single interface validation is removed.

Signed-off-by: Edward Haas <edwardh@redhat.com>