kubvirt: Fix aarch64 machine type and OVMF path (#2863)
* kubvirt: Fix aarch64 OVMF path

As documented [1] downstream aarch64 guests require a different UEFI
firmware to x86_64. This is provided by the edk2-aarch64 package with
the resulting firmware placed under /usr/share/AAVMF.

This change updates the OVMFPath configured for use with such guests on
aarch64 and also updates some tests to assert this is set.

[1] https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/9/html-single/configuring_and_managing_virtualization/index#how-virtualization-on-arm-64-differs-from-amd64-and-intel64_feature-support-and-limitations-in-rhel-9-virtualization

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>

* kubevirt: Introduce separate arch machine type env variables

This change introduces arch specific versions of the machine type env
variable. This avoids the same machine type being set for both x86_64
and aarch64. The original `MACHINETYPE` env variable is supported for a
release to ensure any downstream consumers continue to work.

We also stop setting the deprecated top level machineType [1]
configurable in favour of these arch specific versions.

[1] https://kubevirt.io/api-reference/main/definitions.html#_v1_kubevirtconfiguration

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>

---------

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>