virt-config: Raise an error if SLIRP is set by default on an interface
Given Kubevirt CR configuration to enable SLIRP and allow it to be set
by default (in case no network interface is defined), return with an
error.

The default setting is performed on a mutating admitter webhook for both
VM and VMI CRs.
In addition, the defaulting is executed when the VM controller prepares
a VMI for creation.
We would expect the failure to be raised at the admitter, failing
immediately the CR creation. However, even if it will fail at the VM
controller, the outcome will be somehow similar: VMI will not be
created, i.e. the VM will not be able to start.

Signed-off-by: Edward Haas <edwardh@redhat.com>