Move vm methods to vm_expansion
Our `vm` struct embed generated `VirtualMachineInterface`.
The latter defines the custom methods but does not
implement them, which are defined in the `vm`.

We can move the expanded methods implementations
in the `virtualmachine_expansion` file so that our `vm`
can simply inherit them.

Also, switched using full `AbsPath` with fmt.Sprintf()
to request composition with `Namespace()`, `Resource()`,
`Name()` and `SubResource()`.

N.B. the `PortForward` method implementation remains in our
`vm` struct because we cannot still implement it in the
expansion file, due to the lack of the clientConfig.

Signed-off-by: fossedihelm <ffossemo@redhat.com>