Move vmi methods to vmi_expansion
Our `vmi` struct embed generated `VirtualMachineInstanceInterface`.
The latter defines the custom methods but does not
implement them, which are defined in the `vm`.

We can move the expanded methods implementations
in the `virtualmachineinstance_expansion` file so that our
`vmi` can simply inherit them.

Also, switched using full `AbsPath` with fmt.Sprintf()
to request composition with `Namespace()`, `Resource()`,
`Name()` and `SubResource()`.

N.B. the methods that return StreamInterface remain in our
`vmi` struct because we cannot still implement it in the
expansion file, due to the lack of the clientConfig.

Signed-off-by: fossedihelm <ffossemo@redhat.com>