rpms: add virtctl-usbredir target with hwdata
hwdata rpm provides metadata that we can use to match device
information such as vendor/product from USB, PCI and others to actual
Vendor and Product names.

This rpm will be needed at build time of virtctl usbredir subcomand as
we are going to embed usb.ids into it, to avoid runtime dependency of
this package.

The `make rpm-deps` happens in the following commit to help with
git rebase whenever needed in the course of the PR.

Signed-off-by: Victor Toso <victortoso@redhat.com>