liveupdate: Add LiveUpdater for LiveUpdateHandlers
This commits adds the `LiveUpdater` which is
used for registering types implementing the LiveUpdateHandler
interface.

Types implementing `LiveUpdateHandler` are used to manage
a subset of live-updatable fields in the VMSpec, they just
have to specify which fields they manage and a liveUpdateHandler
implementation to react of changes to those fields.

This will help clean up the current implementation of the
vmRolloutStrategy feature and should make it easier to introduce
new live-updatable fields.

Unit tests are also added.

Signed-off-by: Antonio Cardace <acardace@redhat.com>