vmi ctrl: convert expectations for pod
The mockinterfaces need to be instructed to what
expect before the controller execution, while using
the fakeclient allow us to check resources after the
controller execution.

Convert the `shouldExpect` functions, which prepare the
mockinterface to `expect` ones, which check the fakeclient.

In particular:
- shouldExpectMatchingPodCreation -> expectMatchingPodCreation
- shouldExpectPodCreation -> expectPodExists
- shouldExpectPodDeletion -> expectPodDoesNotExist
- dropped shouldExpectMultiplePodDeletions
- shouldExpectHotplugPod -> expectHotplugPod

N.B. shouldExpectPodDeletion was also moved under a
specific `DescribeTable` which requires it.
This will be address in the future.

Signed-off-by: fossedihelm <ffossemo@redhat.com>