vmi ctrl: fix pod annotation patch tests
There are two tests that want to check the
existence(or not) of specific network annotations
on the pod.

Currently, those tests are not checking correctly
the annotations, since they are looking at the
`pod` variable that is defined under the test scope itself.

What we want to check is that the `updatedPod` (IOW: in the
fakeclient) has the correct annotations.

Signed-off-by: fossedihelm <ffossemo@redhat.com>

(cherry picked from commit e573666e3947e9190675e4c7f94eeef8035e40f6)
Signed-off-by: fossedihelm <ffossemo@redhat.com>