fix(vmi ctrl): remove empty activepods and volumestatus
`NewPendingVirtualMachine` function returns a vmi in
pending status. This vmi is used in the tests and it
is registered in the fakeclient and the informers.

Both activePods and volumeStatus are defined as `json:<name,omitempty>`.
This means that the in real clusters it is not possible that
there will be a Pod with empty activePods (because the marshall will
convert it to nil). However, currently, we are storing a vmi with
empty (non-nil) activepods in the vmi informer.
Since the pod in the vmi informer is used by the vmi controller
to build the patch operations, the empty map causes a failure in the
`test` patch operation in the test `should add active pods to status if VMI is in running state`.

In other words, the resource stored in the vmi fakeclient is different to
what we put in the vmi informer because in the fakeclient there is
a nil value, while in the vmi informer an empty map.

The empty map will result in this patch operation:
`{ "op": "test", "path": "/status/activePods", "value": {} }`
that fails because it should be:
`{ "op": "test", "path": "/status/activePods", "value": null }`

This is a clear example of how the fakeclient help us to create
tests closer to reality, which results in an increasing of
stability.

Even if no failures was observed around the VolumeStatus,
the same logic can be applied, so here we are going to
align its behavior too.

Signed-off-by: fossedihelm <ffossemo@redhat.com>