vmi ctrl: fix vmi status hotplug disk tests
There are two tests that checks the vmi phase and volumestatus.

Regarding the phase, currently, those tests are not performing
the right check, since they are looking at the
`vmi` variable that is defined under the test scope itself.

What we want to check is that the `updatedVmi` (IOW: in the
fakeclient) has the correct phase.

Regarding the volumestatus, we can switch checking the
updated vmi from the fakeclient instead of expexting the
right patch operation.

Signed-off-by: fossedihelm <ffossemo@redhat.com>
(cherry picked from commit e66a79edfde98f61f0521c9e621ed16114846ba8)