fix: do not override original vmi during eventually cycle
The `watchVMIForPhase` function tries to fetch the vmi
waiting for a specific phase.

During the `Eventually` the original vmi variable is
continuously overwritten by the retrieved value one.

This could lead to a failure of the whole `Eventually`
in case of one of the `Get` before reaching the desired phase fails.
This is because the later attempts will not provide to the
`Get` operation the correct vmi name, Actually the later attempts
will provide an empty name.
In fact, a `Get` always returns a vmi (with empty fields in case of error).

Let's store the retrieved vmi in a separate variable.

Failure example: https://prow.ci.kubevirt.io/view/gs/kubevirt-prow/pr-logs/pull/kubevirt_kubevirt/11685/pull-kubevirt-e2e-k8s-1.28-sig-storage/1781258067900370945

Signed-off-by: fossedihelm <ffossemo@redhat.com>