tests: hide Copy/DeleteAlpineWithNonQEMUPermissions
CopyAlpineWithNonQEMUPermissions and its sibling
DeleteAlpineWithNonQEMUPermissions do a complex preparation (and later
cleanup) of an Alpine test image for NFS tests.

I do not understand the tests that use them and why
CopyAlpineWithNonQEMUPermissions is called ForEach of them. Cleaning up
this is left as a future excercise.

This commit only moves the convoluted functions next to the tests that
uses them.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>