refactoring(vmim ctrl): use fakeclient
Register resources to the fakeclient during their
addition to the informers. In this way wa can interact
directly with the fakeclient that will store the
changes that will be applied to the resources.

Created `addPod`convenience function and removed
`addMigrationPolicy` since `addMigrationPolicies`
can be used.

These will be used in subsequent commits.

Signed-off-by: fossedihelm <ffossemo@redhat.com>