watch: fix live updateable volumes
This PR fixes the live-updatabe volumes. Without this change, all kind of
volumes and disks are updatable. However, this is a wrong assumption.
The only volumes and disk that can change while the VM is running are the
hotpluggable volumes, for the other, the restart requiered condition needs
to be added.

Signed-off-by: Alice Frosi <afrosi@redhat.com>