fix(vmim ctrl): introduce addNodeNameToVMI
In real clusters when the vmi status has the `NodeName`
valued, the vmi is also labelled with
"virtv1.NodeNameLabel: nodeName" label.

The missing label in the unit tests results, sometimes, in a
failure in the patch composition since the resource stored
in the vmi fakeclient is different to what we put in the
vmi informer.

Signed-off-by: fossedihelm <ffossemo@redhat.com>