vmim ctrl: convert expectations for vmi
The mockinterfaces need to be instructed to what
expect before the controller execution, while using
the fakeclient allow us to check resources after the
controller execution.

Convert the `shouldExpect` functions, which prepare the
mockinterface to `expect` ones, which check the fakeclient.

In particular, generic shouldExpectVirtualMachineInstancePatch has been
converted to splitted functions:
- expectVirtualMachineInstanceMigrationState
- expectVirtualMachineInstanceLabels
- expectVirtualMachineInstanceMigrationConfiguration

Signed-off-by: fossedihelm <ffossemo@redhat.com>