fix(vmim ctrl): improve migration policy tests
In the `DescribeTable` "should override cluster-wide migration configurations when"
the vmi stored in the fakeclient differs from the one used to
check the generated patch.
This because the vmi is registered in the fakelient in the BeforeEach,
but later, in the body it is modified with `generatePolicyAndAlignVMI`.
This is wrong and results in error using the fakeclient.

Signed-off-by: fossedihelm <ffossemo@redhat.com>