drop copyOriginalCDI
operator.go is a single file in its package. There is no reason to use
lambdas here, because there is no risk for naming pollution, as there
are no other files in this package.

copyOriginalCDI is a lambda function that is only called once. It's also
not very large function, so this commit just inline its content in the
only place it was called from.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>