change allPodsAreReady to be a regular function
Also, rename the function to `allKvInfraPodsAreReady`.

operator.go is a single file in its package. There is no reason to use
lambdas here, because there is no risk for naming pollution, as there
are no other files in this package.

This commit changes allPodsAreReady lambda, to be a regular function.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>