drop several patch lambdas
operator.go is a single file in its package. There is no reason to use
lambdas here, because there is no risk for naming pollution, as there
are no other files in this package.

In addition to be lambdas, the following functions only differ by the
json data used to patch the KubeVirt CR. All the patch logic is common.

* patchKvProductNameVersionAndComponent
* patchKvVersionAndRegistry
* patchKvVersion
* patchKvCertConfig

This commit adds the new `patchKV` function to do the common logic. The
building of the json data + calling to `patchKV` is now done inline.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>