virtctl: usbredir: provide vendor/product from USB
The hwdata rpm provides metadata files that we can use to identify
vendors and products. The hwdata/usb.ids file provides such metadata
for USB devices and this patch adds a parser and related logic to
fetch it.

We embed this file into usbredir package to not have runtime
dependency on it. The rpm used was introduced in previous commit.

Signed-off-by: Victor Toso <victortoso@redhat.com>