instancetype, tests: Use Alpine for memory overcommit test
The guestless VMI provides 1M of memory that is copied into the instance
type used for this test. As the instance type also enables memory
overcommit we actually end up with less than 1M requested within the VMI
spec of the VM. The simple fix here is to switch to another test VMI
that provides more memory, in this case Alpine.

The use of preferences is also dropped from this test as they are
unrelated to the feature.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>