UserDataSecretRef was incorrectly coding to secretRef
Convention for CloudInitNoCloudSource has the general form of:
```
UserData -> `userData`
NetworkData -> `networkData`
```
But the UserDataSecretRef is miscoded to just `secretRef`. Currently:
```
UserDataSecretRef -> `secretRef`
NetworkDataSecretRef -> `networkDataSecretRef`
```

This is inconsistent and is creating confusion in the documentation. [Here the secret is described](https://github.com/kubevirt/kubevirt/blob/main/docs/cloud-init.md) as `userDataSecretRef` which WILL NOT WORK until this PR is accepted. In [other documentation](https://kubevirt.io/user-guide/virtual_machines/startup_scripts/#cloud-init-userdata-as-k8s-secret) the current operation is described correctly, but it is inconsistent with the rest of the CloudInitNoCloudSource naming convention.

Suggest this PR be accepted and the documentation in the User Guide be updated to userDataSecretRef instead of secretRef.

Signed-off-by: Jonathan Siegel <248302+usiegj00@users.noreply.github.com>