virtwrap: change containerdisks handling
Add new behavior and handling for containerdisk. In the previous
version, containerdisk artifacts were mounted by virt-handler and
virt-launcher needed to synchronize before being able to access them.

With the new support, the coordination mechanism has been completly drop
and virt-launcher needs simply to create the overlay images and the
symlinks for the expected locations.

Signed-off-by: Alice Frosi <afrosi@redhat.com>