node_labeller.go: Fix test_id: 6247 We are setting kubevirt.configuration.ObsoleteCPUModels to nil to align with test expectations. The test assumes that kubevirt.configuration.ObsoleteCPUModels is not set.
Before this PR, if kubevirt.configuration.ObsoleteCPUModels
was set before the tests started running,
then the test would fail.

After this PR, if kubevirt.configuration.ObsoleteCPUModels
was set before the tests started running, it will be set
to nil to prevent failure.

Additionally, we will wait for 30 seconds before failing
to give the node labeller enough time to set the new
labels on the nodes.

Signed-off-by: bmordeha <bmordeha@redhat.com>