refactoring(vmirs ctrl): use fakeclient
Register resources to the fakeclient during their
addition to the informers. In this way we can interact
directly with the fakeclient that will store the
changes that will be applied to the resources.

Created `addVMI`convenience function.
This allows us to remove some fake prepend reactor, since
the resources are correctly registered (or not), in the client.

Signed-off-by: fossedihelm <ffossemo@redhat.com>