tests, libsecret: Generalize data used for secret
Users can pass in data as a map of strings or a map of bytes.

The current usage requires only a map of strings, but the bytes version
has been created to allow passing in also that type when needed.

In this form, it should be easier for the callers to pass the data in
the most convenient manner.

Signed-off-by: Edward Haas <edwardh@redhat.com>