VM Admitter: Block CPU Hotunplug
When the socket count of a VM is changed the following steps are
performed:

1. The VirtualMachineInstance is updated.
2. A migration is triggered.
3. The target virt-launcher pod is created according to the updated VMI.
4. The domain is migrated as-is.
5. On the target virt-launcher pod - the domain is updated to fit.

This works well for CPU hotplug, but is problematic for CPU hotuplug
is several scenarios, for example:
1. When using dedicated CPUs
2. When using `networkInterfaceMultiqueue: true`

In these cases, the migration fails due to incompatibilities between
the source and target virt-launcher pods.

Block CPU hotunplug until this issue will be resolved in the future.

Signed-off-by: Orel Misan <omisan@redhat.com>