Ctrl, VM: Require restart when socket count is reduced
Currently, it is possible to reduce the sockets count
of a VirtualMachine object.
Doing so causes the following steps to happen:

1. The VirtualMachineInstance is updated.
2. A migration is triggered.
3. The target virt-launcher pod is created according to the updated VMI.
4. The domain is migrated as-is.
5. On the target virt-launcher pod - the domain is updated to fit.

This works well for CPU hotplug, but is problematic for CPU hotunplug in
several scenarios, for example:
1. When using dedicated CPUs
2. When using `networkInterfaceMultiqueue: true`

Add the RestartRequired condition on the VM in case the CPU socket
count is reduced.
This will make sure the virt-launcher pod and the domain are aligned.

Signed-off-by: Orel Misan <omisan@redhat.com>
(cherry picked from commit 000c9c3e82c20f2577b3008bef1e6c22d64deb64)
Signed-off-by: fossedihelm <ffossemo@redhat.com>