virt-config: Fix discontinued FG reporting query
When querying a FG if it is enabled or not, keep on reporting the actual
FG setting per the configuration except the case in which the FG is GA.

Specifically, the current logic returned `false` for discontinued FG,
which may cause existing running VMs to fail validation.

Therefore, make sure that even for discontinued FG the query depends on
the configuration (and is not statically `false`).

Signed-off-by: Edward Haas <edwardh@redhat.com>