feature-gate, deprecation: Add FG deprecation validation for macvtap
Validate discontinued feature-gates at the relevant admitters for new
object creation.

Updates are not validated in order to allow existing workloads to
continue working.

When a discontinued FG is detected on a resource creation operation, an
error cause is reflected to the user.

At the moment, the `macvtap` network core binding is marked as
discontinued.

Signed-off-by: Edward Haas <edwardh@redhat.com>