controller: Apply preferences to hot plugged interfaces
Previously any preferences referenced by the VirtualMachine would not be
applied to hot plugged interfaces.

This change adds a crude check for new interfaces after
ApplyDynamicIfaceRequestOnVMI has been called before then calling
applyDevicePreferences to lookup and apply any associated preferences to
a copy of the VMI that is later used to patch the VMI object itself.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>