fix(e2e postcopy): Create VM with RerunOnFailure
Currently, postcopy restart test relies on the fact
that when adding `RerunOnFailure` runStrategy to the vm
spec, the vm is automatically started.

This is not true, since the `RerunOnFailure` will start
the vm only if the latter is directly created with it.
Adding it will not act as a start request.

This bug was exposed by previous commits, since some status
updates were overwritten by the update.

Signed-off-by: fossedihelm <ffossemo@redhat.com>