e2e,compute: Quarantine flaky force restart e2e test
This test is very flaky at the moment[1] and meets the criteria for
quarantine.

The large timeout[2] in this test allowed the test VM to gracefully
shutdown so it may not have been testing force restart at all.

This flake could be exposing an issue with force restart.

Quarantining for now to allow for investigation.

[1] https://search.ci.kubevirt.io/?search=test_id%3A3007&maxAge=48h&context=1&type=bug%2Bissue%2Bjunit&name=&excludeName=&maxMatches=5&maxBytes=20971520&groupBy=job
[2] https://github.com/kubevirt/kubevirt/blob/037618edc09d7a490f9a6bfed6b60ed31339600c/tests/vm_test.go#L791

Signed-off-by: Brian Carey <bcarey@redhat.com>