virt-controller, hot plug: Avoid race between status update and annotation update
Determining if an annotation update is required is done according to the
interfaces in the VMI status.
Currently, updating the VMI status with multus infoSource happens
after the first annotation update attempt.
This causes an unrequired attempt to update annotations,
because the status is completely missing at this stage.

Fix it by updating annotations only after the VMI is running,
eliminating the transient state false hotplug detection.

Signed-off-by: Or Shoval <oshoval@redhat.com>