instancetype: Add missing legacy v1beta1 CR upgrade test cases
e0ee8356431d69b started to label CRs with the version of the stored
object but this was sometime after v1beta1 was introduced. As such there
might be v1beta1 CRs in the wild without this label.

This change introduces test cases covering this scenario, dropping the
label from the CR *and* importantly using a different name for the
resource.

We need to change the name of the CR created by CreateControllerRevision
to fully test this flow as GetRevisionName was recently changed to
include the version of the object. As such we expect that we can create
a new CR when upgrading any legacy CR without the version label.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>