virt-controller: Create ipam claims for new pods
Once a new pod is created, create the required ipam claims.
Note we don't support hotplug yet, so it the new method
can support only hashed interfaces naming scheme.

We are using a builder method WithIPAMClaimManager,
this way we don't affect the non relevant unit tests,
so we don't need to mock anything.

Signed-off-by: Or Shoval <oshoval@redhat.com>