instancetype: Enforce preference arch requirements
This change starts to enforce any architecture requirements provided by
a preference when an architecture is provided by the provided
VirtualMachineInstanceSpec.

It's worth noting that an architecture is always provided in our initial
call to CheckPreferenceRequirements in the VirtualMachine validation
webhook as defaults are applied prior to this. These changes are later
persisted by the mutation webhook if not already present in the
VirtualMachine.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>