virt-controller, hotplug: Block hotplug on VM level for persistent IP interfaces
Since we don't support yet persistent IPs hotplug / hotunplug,
block it on VM level, and log about the blocking.
This way the VMI will still reflect the right state as done for
interfaces with ordinal naming scheme.

On the other hand, we must pass the ipam claim parameters,
when recalculating the annotations upon hotplug / hotunplug,
so the ipam claim reference won't be stripped, once other
interface is hotplugged / hotunplugged.

Signed-off-by: Or Shoval <oshoval@redhat.com>