virt-operator/generate: Add node-restriction policy
In order to restrict node patch of non-kubevirt labels and annotations,
Introducing new validatingAdmissionPolicy and
validatingAdmissionPolicyBinding objects that target the virt-handler
user.
non-kubevirt labels and annotations is defined by labels and annotations
that contain "kubevirt.io" string, or "cpu-manager" label which is the
exception.
In future commits these objects will be deployed by virt-operator.

Signed-off-by: Ram Lavi <ralavi@redhat.com>
Co-authored-by: fossedihelm <ffossemo@redhat.com>