network: Remove core passt binding
Processing of passt core binding is removed in a way existing running VM/VMIs
are not disrupted.

New VM/VMIs which are configured to use the passt core binding will fail to setup
the vNIC.

A follow up change will block such a spec at the creation validation
admitter to assure early and stable failure for new VM/VMIs.

Signed-off-by: Or Mergi <ormergi@redhat.com>