virt-config, feature-gates: Control the addition of FGs to the FG list
A static FG list has been maintained, defining FGs directly into the
list upon definition.

This approach has several disadvantages:
- There is no strict control on adding a new FG.
  E.g. an addition of two FGs with the same name is possible.
- FG have to be defined in a central place and cannot be distributed to
  the location that is owned by a feature.

Therefore, the current implementation is refactored to use a
registration function to add FGs in a controlled manner.
Duplications is avoided by using a map instead of a list.

When `RegisterFeatureGate` is called, the inputted FG is either added to
the FG list or overriding an existing FG entry.

Signed-off-by: Edward Haas <edwardh@redhat.com>