tests: remove partial downgrade of KubeVirt
Simplify the operator tests by eliminating the partial downgrade of
KubeVirt.
Downgrades aren't officially supported, however, the operator tests
install the previous version of KubeVirt in order to, then, test the
upgrades. As part of the test preparation, we pass of the latest
KubeVirt version to the previous release.

In the case of the installation during KubeVirt CR, the downgrade was
partial as the operator wasn't delete. This change completly remove the
latest kubevirt version, in this way, we can simulate a proper upgrade
from the previous release.

Signed-off-by: Alice Frosi <afrosi@redhat.com>