virt-controller, vm: Do not fail when processing net hotplug with no pod
When network hotplug is processed at the VM controller, there is a check
that attempts to identify if the touched VM is an old running one with
ordered network interface names. This check is assuming a running pod
exists, but this is not assured to be the case (yet).

Therefore, in order to avoid early failure on the missing pod, just
assume the most conservative option, which is that the VM is using
ordered network interface.
In practice, this will only affect the hot-unplug scenario, by not
reflecting the removal to the VMI in the period the pod is not observed.
Once the pod is observed, it will be processed per the actual state.

This issue has been originally detected by observing the logs reporting
at high rate the error: "Failed to find VMI pod in cache.".

Signed-off-by: Edward Haas <edwardh@redhat.com>