ctrl: fix starting VM with Manual RunStrategy
check for the returned error as the `vmi` variable
being not nil does not necessarily mean that the VMI
actually exists.

Because of this bug if you had a VM with `Manual` RunStrategy
and you were to start the VM while the DataVolumes were not ready
the just issued StartRequest would get trimmed by the virt-controller.
To workaround this you'd have to issue a second start request when the
DVs would get ready.

Unfortunately, our production and test code differ in this
behavior as the production code always returns `vmi != nil` and
the fake client returns the `vmi == nil` if the VMI does not exist.

Signed-off-by: Antonio Cardace <acardace@redhat.com>
(cherry picked from commit f60eab4e87ed9b29d25822a45394c5de4e94bbb3)