virt-api, VMI mutator: Use cached KubeVirt ServiceAccounts
Currently, the VMI mutator webhook checks whether a request had originated
from one of KubeVirt's ServiceAccounts:
1. virt-api
2. virt-controller
3. virt-handler

This is currently done using the `IsKubeVirtServiceAccount` [1] utility function.

The ServiceAccounts' names are a function of the namespace
on which KubeVirt is deployed, and have the following form:
system:serviceaccount:<namespace>:<name>

`IsKubeVirtServiceAccount` reads a file [2] from the filesystem,
in order to get the namespace on which KubeVirt is deployed.

Since this function is called very frequently - this is inefficient.

Create a set of KubeVirt ServiceAccounts on virt-api's startup
and use it in the VMI mutator.

Test that all three KubeVirt serviceAccounts are able to modify
the VMI.

[1] https://github.com/kubevirt/kubevirt/blob/51dff11afcf5df674da504080301881eea97abba/pkg/virt-api/webhooks/utils.go#L83
[2] https://kubernetes.io/docs/tasks/run-application/access-api-from-pod/#directly-accessing-the-rest-api

Signed-off-by: Orel Misan <omisan@redhat.com>