OWNERS: Allow test, compute and network SIGs to own `.golangci.yml`
The golangci linters are managed through the `.golangci.yml`
configuration file.

At the moment, most of the code that is covered by the linter is under
the sig-test, sig-compute and sig-network ownership, therefore it would
be useful if the approvers on those SIGs will also have "approve" access
to accept changes to it.

Signed-off-by: Edward Haas <edwardh@redhat.com>