release, verification: add source tarball signature upload
Adds a step to the release creation that tries to download the
detached signature file - if it can be downloaded, it downloads
the source tarball and verifies against the local id.

If either check fails, it downloads the source tarball from the
target release, generates a detached signature file, and uploads
the signature file to the target release.

The script that provides this functionality has been separated by
intent, since it might be used standalone to attach signatures to
existing releases.

Requires `gh` cli tool.

The user-id to sign with is set via env GPG_USER_ID.

For successful verification on a target system the public key for the
user-id used to create the signature has to be imported first.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>