VM unit: Remove vmi informer
There is no real reason to use informers for
unit tests other than to test callbacks. However
these callbacks can be tested separatly.

By removing the informer we achive simpler tests
as we do not require synchronization anymore.
This has also positive impact on test runtime.

This removes syncCache as it is not needed anymore.

Signed-off-by: Luboslav Pivarc <lpivarc@redhat.com>