tests, net, hotplug: Drop reset-after-hotplug test
The hotplug of bridge binding network interface has been covering a
scenario in which after a successful hotplug, the VM is restarted and
the interfaces are re-examined for existence.

This change drops the 2 tests that cover the described scenario.
It is trivial that once the VM spec has been modified, restarting or
starting (post stopping) a VM is identical to the scenario that does not
involve hotplug at all. I.e. it does not matter why the VM spec changed,
the new created VMI spec is going to be based on the VM spec.

Signed-off-by: Edward Haas <edwardh@redhat.com>