tests, net, hotplug: Drop stopped VM test
The scenario of marking a network interface to be unplugged (set to
`absent` state) can be considered redundant because it is not a
hotunplug (the VM is not running, i.e. no VMI) scenario.

Dropping this test from the e2e tests and depending on a unit test [1]
is more appropriate.

[1] https://github.com/kubevirt/kubevirt/blob/31b146a2a7dbb11afa6d041d3e4d9b1efa484f3c/pkg/virt-controller/network/vmcontroller_test.go#L55

Signed-off-by: Edward Haas <edwardh@redhat.com>